% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_estimate_R.R
\name{exponential_estimate_R}
\alias{exponential_estimate_R}
\title{Bayesian exponential model using stan}
\usage{
exponential_estimate_R(
  fit_exponential,
  n_mean = 2.29,
  b_mean = 0.36,
  label = "label"
)
}
\arguments{
\item{fit_exponential}{a model fit using reacttemporal2::stan_exp_model() or reactemporal2::stan_exp_model_phe()}

\item{n_mean}{shape parameter for inverse gamma serial interval (default from Bi et al)}

\item{b_mean}{rate parameter for inverse gamma serial interval (default from Bi et al)}
}
\value{
An object of class \code{dataframe} with calculations of growth rate, R, doubling/halving time with 95\% CIs and the probability R>1

95\% CI for r and median
}
\description{
Bayesian exponential model using stan
}
