% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_example_data}
\alias{load_example_data}
\title{Load and format example data}
\usage{
load_example_data()
}
\value{
An object of class \code{dataframe} with calculations of growth rate, R, doubling/halving time with 95\% CIs and the probability R>1
}
\description{
Load and format example data
}
