% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_phe_data.R
\name{load_example_phe_data}
\alias{load_example_phe_data}
\title{Load and format example data}
\usage{
load_example_phe_data()
}
\value{
PHE Pillar1&2 case numbers data for England only.
}
\description{
Load and format example data
}
