% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_p_spline_minimum_density.r
\name{plot_p_spline_minimum_density}
\alias{plot_p_spline_minimum_density}
\title{Plotting function for the P_spline model}
\usage{
plot_p_spline_minimum_density(
  X,
  p_spline_fit,
  target_dist_between_knots = 5,
  spline_degree = 3
)
}
\arguments{
\item{X}{date vector.}

\item{target_dist_between_knots}{sets the number of days between adjacent knots (default = 5)}

\item{spline_degree}{sets the degree of the splines (default = 3)}

\item{p_splinefit}{fit of the model to the same set of X values using reacttemporal2::stan_p_spline() or reacttemporal2::stan_p_spline_phe()}
}
\value{
A list of the created plot, and the list of values used in the histogram.

Plot of the model fit
}
\description{
Plotting function for the P_spline model
}
