% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_p_spline_phe.R
\name{plot_p_spline_phe}
\alias{plot_p_spline_phe}
\title{Plotting function for the P_spline model}
\usage{
plot_p_spline_phe(
  X,
  Y,
  p_spline_fit,
  target_dist_between_knots = 5,
  spline_degree = 3,
  ylim = 5000
)
}
\arguments{
\item{X}{date vector.}

\item{Y}{Numeric vector of number of positive samples}

\item{target_dist_between_knots}{sets the number of days between adjacent knots (default = 5)}

\item{spline_degree}{sets the degree of the splines (default = 3)}

\item{ylim}{sets the ylimit of the plot created}

\item{p_splinefit}{fit of the model to the same set o data using reactidd::stan_p_spline_phe()}
}
\value{
A list of the created plot, the raw data and CI's used in the plot, the raw data for the model fit in the plot.

Plot of the model fit
}
\description{
Plotting function for the P_spline model
}
