% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacttemporal2-package.R
\docType{package}
\name{reacttemporal2-package}
\alias{reacttemporal2-package}
\alias{reacttemporal2}
\title{The 'reacttemporal2' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.1. https://mc-stan.org
}
