% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_exp_model_phe.R
\name{stan_exp_model_phe}
\alias{stan_exp_model_phe}
\title{Bayesian exponential model using stan}
\usage{
stan_exp_model_phe(X, Y, N, iter = 5000, warmup = 1000, cores = 1)
}
\arguments{
\item{X}{date vector.}

\item{Y}{Numeric vector of number of cases}

\item{...}{Arguments passed to \code{rstan::sampling} (iter, warmup).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}

Load and run stan model
}
\description{
Bayesian exponential model using stan
}
