% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_data.R
\name{load_example_data}
\alias{load_example_data}
\title{Load and format example data}
\usage{
load_example_data()
}
\value{
An object of class \code{list} with length 2. The first item is a dataframe of number of positive swabs from the REACT data set by region and time. The second item is the same but for total number of swab tests.
}
\description{
Load and format example data
}
